+++ domains/transfer/Transfer.API/API.UnitTests/Transfer/Api/Controllers/TransferAgreementsControllerTests.cs
@@ -2,12 +2,12 @@ using System;
 using System.Collections.Generic;
 using System.Security.Claims;
 using System.Threading.Tasks;
-using API.Shared.Services;
 using API.Transfer.Api.Controllers;
 using API.Transfer.Api.Dto.Requests;
 using API.Transfer.Api.Dto.Responses;
 using API.Transfer.Api.Models;
 using API.Transfer.Api.Repository;
+using API.Transfer.Api.Services;
 using FluentAssertions;
 using FluentValidation;
 using Microsoft.AspNetCore.Http;
