+++ domains/transfer/Transfer.API/API.UnitTests/Transfer/TransferAgreementsAutomation/Service/TransferAgreementsAutomationServiceTest.cs
@@ -2,9 +2,9 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using API.Shared.Services;
 using API.Transfer.Api.Models;
 using API.Transfer.Api.Repository;
+using API.Transfer.Api.Services;
 using API.Transfer.TransferAgreementsAutomation;
 using API.Transfer.TransferAgreementsAutomation.Metrics;
 using API.Transfer.TransferAgreementsAutomation.Service;
