+++ domains/transfer/Transfer.API/API/Claiming/Automation/Services/IProjectOriginWalletService.cs
@@ -1,0 +1,12 @@
+using Google.Protobuf.Collections;
+using ProjectOrigin.WalletSystem.V1;
+using System.Threading.Tasks;
+using System;
+
+namespace API.Claiming.Automation.Services;
+
+public interface IProjectOriginWalletService
+{
+    Task<RepeatedField<GranularCertificate>> GetGranularCertificates(Guid subjectId);
+    Task ClaimCertificate(Guid ownerId, GranularCertificate consumptionCertificate, GranularCertificate productionCertificate, uint quantity);
+}
