+++ domains/transfer/Transfer.API/API/Claiming/Startup.cs
@@ -1,5 +1,6 @@
 using API.Claiming.Api.Repositories;
 using API.Claiming.Automation;
+using API.Claiming.Automation.Services;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace API.Claiming;
@@ -11,6 +12,7 @@ public static class Startup
         services.AddScoped<IClaimRepository, ClaimRepository>();
 
         services.AddScoped<IClaimService, ClaimService>();
+        services.AddScoped<IProjectOriginWalletService, ProjectOriginWalletService>();
 
         services.AddHostedService<ClaimWorker>();
     }
