+++ domains/transfer/Transfer.API/API/Transfer/Api/Controllers/TransferAgreementsController.cs
@@ -3,11 +3,11 @@ using System.Linq;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
 using API.Shared.Extensions;
-using API.Shared.Services;
 using API.Transfer.Api.Dto.Requests;
 using API.Transfer.Api.Dto.Responses;
 using API.Transfer.Api.Models;
 using API.Transfer.Api.Repository;
+using API.Transfer.Api.Services;
 using FluentValidation;
 using FluentValidation.AspNetCore;
 using Microsoft.AspNetCore.Authorization;
