+++ domains/transfer/Transfer.API/API/Transfer/Startup.cs
@@ -2,10 +2,10 @@ using System;
 using System.Linq;
 using System.Text.Json.Serialization;
 using API.Shared;
-using API.Shared.Services;
 using API.Transfer.Api.Models;
 using API.Transfer.Api.Options;
 using API.Transfer.Api.Repository;
+using API.Transfer.Api.Services;
 using API.Transfer.TransferAgreementsAutomation;
 using API.Transfer.TransferAgreementsAutomation.Metrics;
 using API.Transfer.TransferAgreementsAutomation.Service;
