+++ domains/transfer/Transfer.API/API/Transfer/TransferAgreementsAutomation/Service/TransferAgreementsAutomationService.cs
@@ -1,8 +1,8 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using API.Shared.Services;
 using API.Transfer.Api.Repository;
+using API.Transfer.Api.Services;
 using API.Transfer.TransferAgreementsAutomation.Metrics;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Logging;
